Moving everything related to UI in mvvm.ui package
