+++ app/src/main/java/com/mn2square/videolistingmvp/mvvm/ui/VideoUserInteraction.java
@@ -1,4 +1,4 @@
-package com.mn2square.videolistingmvp.mvvm;
+package com.mn2square.videolistingmvp.mvvm.ui;
 
 /**
  * Created by nitinagarwal on 4/9/17.
