+++ app/src/main/java/com/mn2square/videolistingmvp/mvvm/ui/savedlistfragment/SavedListFragmentImpl.java
@@ -1,13 +1,13 @@
-package com.mn2square.videolistingmvp.mvvm.swipetabfragments.savedlistfragment;
+package com.mn2square.videolistingmvp.mvvm.ui.savedlistfragment;
 
 import com.github.ksoichiro.android.observablescrollview.ObservableListView;
 import com.github.ksoichiro.android.observablescrollview.ObservableScrollViewCallbacks;
 import com.github.ksoichiro.android.observablescrollview.ScrollState;
 import com.mn2square.videolistingmvp.R;
-import com.mn2square.videolistingmvp.mvvm.VideoUserInteraction;
+import com.mn2square.videolistingmvp.mvvm.ui.VideoUserInteraction;
 import com.mn2square.videolistingmvp.mvvm.pojo.VideoListInfo;
-import com.mn2square.videolistingmvp.mvvm.MvvmVideoListActivity;
-import com.mn2square.videolistingmvp.mvvm.swipetabfragments.adapters.VideoListAdapter;
+import com.mn2square.videolistingmvp.mvvm.ui.MvvmVideoListActivity;
+import com.mn2square.videolistingmvp.mvvm.ui.adapters.VideoListAdapter;
 
 import android.arch.lifecycle.Observer;
 import android.arch.lifecycle.ViewModelProviders;
