+++ app/src/main/java/com/mn2square/videolistingmvp/mvvm/ui/savedlistfragment/SavedListViewModel.java
@@ -1,4 +1,4 @@
-package com.mn2square.videolistingmvp.mvvm.swipetabfragments.savedlistfragment;
+package com.mn2square.videolistingmvp.mvvm.ui.savedlistfragment;
 
 import com.mn2square.videolistingmvp.mvvm.pojo.VideoListInfo;
 import com.mn2square.videolistingmvp.mvvm.repository.VideoListRepositoryImpl;
